+++ inst/NEWS.md
@@ -4,9 +4,12 @@ __MEDseq: Mixtures of Exponential-Distance Models with Covariates__
 ### Improvements, Bug Fixes & Miscellaneous Edits
 * Function `seqdef` added as an exact copy of `TraMineR::seqdef`, to enable experienced  
   users of `MEDseq` & `TraMineR` to use the former without needing to explicitly load the latter.
+* The `plot.MEDseq` function's `type` arg. gains the option `"dH"`,  
+provided version `2.2-4` or later of the `TraMineR` package is installed.
 * `wKModes` now also returns `x$tot.withindiff` (i.e. `sum(x$withindiff)`).
-* Fixed rare bug in tie-breaking for modal sequence estimate.
+* Fixed rare bugs in tie-breaking for modal sequence estimate.
 * New function `MEDseq_AvePP` added.
+* Fixed documentation typos.
 
 ## MEDseq v1.3.3 - (_10<sup>th</sup> release [patch update]: 2022-03-28_)
 ### Improvements, Bug Fixes & Miscellaneous Edits